moved the license text out to a separate js file;
